# services/auth/api/openapi3/openapi.yaml
openapi: 3.0.3

info:
  title: Traxex Auth API
  version: 1.0.0

paths:
  /auth/register:
    post:
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: email token

  /auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          description: invalid credentials

  /auth/logout:
    post:
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        400:
          description: bad request
        404:
          description: not found
        500:
          description: internal server error
        503:
          description: server unavailable
    
  /auth/validate-token:
    post:
      operationId: validate token
      requestBody:
        required: true
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/ValidateTokenRequest'
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
        401:
          description: invalid or expired token
        500:
          description: internal server error
  
  /auth/refresh:
    post:
      operationId: refresh token
      requestBody:
        required: true
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          description: invalid or expired token
        500:
          description: internal server error

components: 
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    ValidateTokenRequest:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
    
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer

    LogoutResponse:
      type: object
      properties:
        success:
          type: boolean

    ValidateTokenResponse:
      type: object
      properties:
        user_id:
          type: string
        
            